backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
django-cors-headers
drf-spectacular
psycopg2-binary
pillow
celery
redis
firebase-admin
